core/requirements.txt
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0
pyyaml>=6.0
numpy>=1.24.0
scikit-learn>=1.2.0
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pytest>=7.0.0
